chore: switch 2 CMake

+++ ch13/barproto.h
@@ -4,6 +4,8 @@
 ** This code is Copyright 1999 by Dann Corbit
 */
 
+#include <stdio.h>
+
 /*
 ** Definitions for our priority queue model merge sort.
 */
@@ -16,12 +18,12 @@ typedef struct tag_fileset {
 
 extern int fgetitem(struct tag_fileset *p);
 extern char *fdeletemin(struct tag_fileset *p, char *end);
-extern void fshellsort(struct tag_fileset *fset, unsigned int count);
-extern void fnormalize(struct tag_fileset *fset, unsigned int count);
+extern void fshellsort(struct tag_fileset *fset, size_t count);
+extern void fnormalize(struct tag_fileset *fset, size_t count);
 extern void foutputr(char *e);
 extern int ge(char *l, char *r);
 extern int le(char *l, char *r);
 extern int lt(char *l, char *r);
-static int readlines(char *file_name, char **lines, int maxlines, unsigned int *offset);
+static int readlines(char *file_name, char **lines, int maxlines, size_t *offset);
 extern void writelines(char **t, int nlines, FILE *fout);
 extern int main(int argc, char **argv);
